Migrate usage of 'index' to 'circuit'
