+++ marlin/src/ahp/indexer.rs
@@ -33,13 +33,13 @@ use snarkvm_utilities::{serialize::*, ToBytes};
 use core::marker::PhantomData;
 use derivative::Derivative;
 
-/// Information about the index, including the field of definition, the number of
+/// Information about the circuit, including the field of definition, the number of
 /// variables, the number of constraints, and the maximum number of non-zero
 /// entries in any of the constraint matrices.
 #[derive(Derivative)]
 #[derivative(Clone(bound = ""), Copy(bound = ""))]
 #[derive(Debug, CanonicalSerialize, CanonicalDeserialize)]
-pub struct IndexInfo<F> {
+pub struct CircuitInfo<F> {
     /// The total number of variables in the constraint system.
     pub num_variables: usize,
     /// The number of constraints.
@@ -51,15 +51,14 @@ pub struct IndexInfo<F> {
     f: PhantomData<F>,
 }
 
-impl<F: PrimeField> IndexInfo<F> {
-    /// The maximum degree of polynomial required to represent this index in the
-    /// the AHP.
+impl<F: PrimeField> CircuitInfo<F> {
+    /// The maximum degree of polynomial required to represent this index in the AHP.
     pub fn max_degree(&self) -> usize {
         AHPForR1CS::<F>::max_degree(self.num_constraints, self.num_variables, self.num_non_zero).unwrap()
     }
 }
 
-impl<F: PrimeField> ToBytes for IndexInfo<F> {
+impl<F: PrimeField> ToBytes for CircuitInfo<F> {
     fn write<W: Write>(&self, mut w: W) -> Result<(), std::io::Error> {
         (self.num_variables as u64).write(&mut w)?;
         (self.num_constraints as u64).write(&mut w)?;
@@ -80,9 +79,9 @@ pub type Matrix<F> = Vec<Vec<(F, usize)>>;
 /// 2) `{a,b,c}` are the matrices defining the R1CS instance
 /// 3) `{a,b,c}_star_arith` are structs containing information about A^*, B^*, and C^*,
 /// which are matrices defined as `M^*(i, j) = M(j, i) * u_H(j, j)`.
-pub struct Index<F: PrimeField> {
+pub struct Circuit<F: PrimeField> {
     /// Information about the index.
-    pub index_info: IndexInfo<F>,
+    pub index_info: CircuitInfo<F>,
 
     /// The A matrix for the R1CS instance
     pub a: Matrix<F>,
@@ -99,7 +98,7 @@ pub struct Index<F: PrimeField> {
     pub c_star_arith: MatrixArithmetization<F>,
 }
 
-impl<F: PrimeField> Index<F> {
+impl<F: PrimeField> Circuit<F> {
     /// The maximum degree required to represent polynomials of this index.
     pub fn max_degree(&self) -> usize {
         self.index_info.max_degree()
@@ -127,7 +126,7 @@ impl<F: PrimeField> Index<F> {
 
 impl<F: PrimeField> AHPForR1CS<F> {
     /// Generate the index for this constraint system.
-    pub fn index<C: ConstraintSynthesizer<F>>(c: &C) -> Result<Index<F>, Error> {
+    pub fn index<C: ConstraintSynthesizer<F>>(c: &C) -> Result<Circuit<F>, Error> {
         let index_time = start_timer!(|| "AHP::Index");
 
         let constraint_time = start_timer!(|| "Generating constraints");
@@ -160,7 +159,7 @@ impl<F: PrimeField> AHPForR1CS<F> {
             return Err(Error::InvalidPublicInputLength);
         }
 
-        let index_info = IndexInfo {
+        let index_info = CircuitInfo {
             num_variables,
             num_constraints,
             num_non_zero,
@@ -191,7 +190,7 @@ impl<F: PrimeField> AHPForR1CS<F> {
         end_timer!(c_arithmetization_time);
 
         end_timer!(index_time);
-        Ok(Index {
+        Ok(Circuit {
             index_info,
 
             a,
