+++ marlin/src/lib.rs
@@ -58,361 +58,61 @@ macro_rules! eprintln {
     ($($arg: tt)*) => {};
 }
 
+/// The Marlin public parameters for a given circuit.
+pub mod circuit_parameters;
+pub use circuit_parameters::*;
+
+/// The Marlin circuit proving key.
+pub mod circuit_proving_key;
+pub use circuit_proving_key::*;
+
+/// The Marlin circuit verifying key.
+pub mod circuit_verifying_key;
+pub use circuit_verifying_key::*;
+
 /// Implements a Fiat-Shamir based Rng that allows one to incrementally update
 /// the seed based on new messages in the proof transcript.
 pub mod rng;
 use rng::FiatShamirRng;
 
-mod error;
-pub use error::*;
-
-mod data_structures;
-pub use data_structures::*;
+mod errors;
+pub use errors::*;
 
 /// Implements an Algebraic Holographic Proof (AHP) for the R1CS indexed relation.
 pub mod ahp;
 pub use ahp::AHPForR1CS;
 
-/// The public parameters used for an instantiation of a circuit.
-pub mod parameters;
-pub use parameters::*;
+/// A generic implementation of the Marlin proof system..
+pub mod marlin;
+pub use marlin::*;
 
 /// The Marlin zkSNARK proof.
 pub mod proof;
 pub use proof::*;
 
 pub mod snark;
+pub use snark::*;
 
 #[cfg(test)]
 mod test;
 
-use crate::ahp::EvaluationsProvider;
-use snarkvm_models::{curves::PrimeField, gadgets::r1cs::ConstraintSynthesizer};
-use snarkvm_polycommit::{Evaluations, LabeledCommitment, PCUniversalParams, PolynomialCommitment};
-use snarkvm_utilities::{bytes::ToBytes, rand::UniformRand, to_bytes};
+use snarkvm_polycommit::PolynomialCommitment;
 
-use core::marker::PhantomData;
-use digest::Digest;
-use rand_core::RngCore;
+/// The universal public parameters for the argument system.
+pub type UniversalSRS<F, PC> = <PC as PolynomialCommitment<F>>::UniversalParams;
 
-/// The compiled argument system.
-pub struct Marlin<F: PrimeField, PC: PolynomialCommitment<F>, D: Digest>(
-    #[doc(hidden)] PhantomData<F>,
-    #[doc(hidden)] PhantomData<PC>,
-    #[doc(hidden)] PhantomData<D>,
-);
-
-impl<F: PrimeField, PC: PolynomialCommitment<F>, D: Digest> Marlin<F, PC, D> {
-    /// The personalization string for this protocol. Used to personalize the
-    /// Fiat-Shamir rng.
-    pub const PROTOCOL_NAME: &'static [u8] = b"MARLIN-2019";
-
-    /// Generate the universal prover and verifier keys for the
-    /// argument system.
-    pub fn universal_setup<R: RngCore>(
-        num_constraints: usize,
-        num_variables: usize,
-        num_non_zero: usize,
-        rng: &mut R,
-    ) -> Result<UniversalSRS<F, PC>, Error<PC::Error>> {
-        let max_degree = AHPForR1CS::<F>::max_degree(num_constraints, num_variables, num_non_zero)?;
-        let setup_time = start_timer!(|| {
-            format!(
-                "Marlin::UniversalSetup with max_degree {}, computed for a maximum of {} constraints, {} vars, {} non_zero",
-                max_degree, num_constraints, num_variables, num_non_zero,
-            )
-        });
-
-        let srs = PC::setup(max_degree, rng).map_err(Error::from_pc_err);
-        end_timer!(setup_time);
-        srs
-    }
-
-    /// Generate the index-specific (i.e., circuit-specific) prover and verifier
-    /// keys. This is a deterministic algorithm that anyone can rerun.
-    #[allow(clippy::type_complexity)]
-    pub fn index<C: ConstraintSynthesizer<F>>(
-        srs: &UniversalSRS<F, PC>,
-        c: &C,
-    ) -> Result<(IndexProverKey<F, PC>, IndexVerifierKey<F, PC>), Error<PC::Error>> {
-        let index_time = start_timer!(|| "Marlin::Index");
-
-        // TODO: Add check that c is in the correct mode.
-        let index = AHPForR1CS::index(c)?;
-        if srs.max_degree() < index.max_degree() {
-            return Err(Error::IndexTooLarge);
-        }
-
-        let coeff_support = AHPForR1CS::get_degree_bounds(&index.index_info);
-        // Marlin only needs degree 2 random polynomials
-        let supported_hiding_bound = 1;
-        let (committer_key, verifier_key) =
-            PC::trim(&srs, index.max_degree(), supported_hiding_bound, Some(&coeff_support))
-                .map_err(Error::from_pc_err)?;
-
-        let commit_time = start_timer!(|| "Commit to index polynomials");
-        let (index_comms, index_comm_rands): (_, _) =
-            PC::commit(&committer_key, index.iter(), None).map_err(Error::from_pc_err)?;
-        end_timer!(commit_time);
-
-        let index_comms = index_comms.into_iter().map(|c| c.commitment().clone()).collect();
-        let index_vk = IndexVerifierKey {
-            index_info: index.index_info,
-            index_comms,
-            verifier_key,
-        };
-
-        let index_pk = IndexProverKey {
-            index,
-            index_comm_rands,
-            index_vk: index_vk.clone(),
-            committer_key,
-        };
-
-        end_timer!(index_time);
-
-        Ok((index_pk, index_vk))
-    }
-
-    /// Create a zkSNARK asserting that the constraint system is satisfied.
-    pub fn prove<C: ConstraintSynthesizer<F>, R: RngCore>(
-        index_pk: &IndexProverKey<F, PC>,
-        c: &C,
-        zk_rng: &mut R,
-    ) -> Result<Proof<F, PC>, Error<PC::Error>> {
-        let prover_time = start_timer!(|| "Marlin::Prover");
-        // Add check that c is in the correct mode.
-
-        let prover_init_state = AHPForR1CS::prover_init(&index_pk.index, c)?;
-        let public_input = prover_init_state.public_input();
-        let mut fs_rng =
-            FiatShamirRng::<D>::from_seed(&to_bytes![&Self::PROTOCOL_NAME, &index_pk.index_vk, &public_input].unwrap());
-
-        // --------------------------------------------------------------------
-        // First round
-
-        let (prover_first_msg, prover_first_oracles, prover_state) =
-            AHPForR1CS::prover_first_round(prover_init_state, zk_rng)?;
-
-        let first_round_comm_time = start_timer!(|| "Committing to first round polys");
-        let (first_comms, first_comm_rands) =
-            PC::commit(&index_pk.committer_key, prover_first_oracles.iter(), Some(zk_rng))
-                .map_err(Error::from_pc_err)?;
-        end_timer!(first_round_comm_time);
-
-        fs_rng.absorb(&to_bytes![first_comms, prover_first_msg].unwrap());
-
-        let (verifier_first_msg, verifier_state) =
-            AHPForR1CS::verifier_first_round(index_pk.index_vk.index_info, &mut fs_rng)?;
-        // --------------------------------------------------------------------
-
-        // --------------------------------------------------------------------
-        // Second round
-
-        let (prover_second_msg, prover_second_oracles, prover_state) =
-            AHPForR1CS::prover_second_round(&verifier_first_msg, prover_state, zk_rng);
-
-        let second_round_comm_time = start_timer!(|| "Committing to second round polys");
-        let (second_comms, second_comm_rands) =
-            PC::commit(&index_pk.committer_key, prover_second_oracles.iter(), Some(zk_rng))
-                .map_err(Error::from_pc_err)?;
-        end_timer!(second_round_comm_time);
-
-        fs_rng.absorb(&to_bytes![second_comms, prover_second_msg].unwrap());
-
-        let (verifier_second_msg, verifier_state) = AHPForR1CS::verifier_second_round(verifier_state, &mut fs_rng);
-        // --------------------------------------------------------------------
-
-        // --------------------------------------------------------------------
-        // Third round
-        let (prover_third_msg, prover_third_oracles) =
-            AHPForR1CS::prover_third_round(&verifier_second_msg, prover_state, zk_rng)?;
-
-        let third_round_comm_time = start_timer!(|| "Committing to third round polys");
-        let (third_comms, third_comm_rands) =
-            PC::commit(&index_pk.committer_key, prover_third_oracles.iter(), Some(zk_rng))
-                .map_err(Error::from_pc_err)?;
-        end_timer!(third_round_comm_time);
-
-        fs_rng.absorb(&to_bytes![third_comms, prover_third_msg].unwrap());
-
-        let verifier_state = AHPForR1CS::verifier_third_round(verifier_state, &mut fs_rng);
-        // --------------------------------------------------------------------
-
-        // Gather prover polynomials in one vector.
-        let polynomials: Vec<_> = index_pk
-            .index
-            .iter()
-            .chain(prover_first_oracles.iter())
-            .chain(prover_second_oracles.iter())
-            .chain(prover_third_oracles.iter())
-            .collect();
-
-        // Gather commitments in one vector.
-        #[rustfmt::skip]
-        let commitments = vec![
-            first_comms.iter().map(|p| p.commitment()).cloned().collect(),
-            second_comms.iter().map(|p| p.commitment()).cloned().collect(),
-            third_comms.iter().map(|p| p.commitment()).cloned().collect(),
-        ];
-        let labeled_comms: Vec<_> = index_pk
-            .index_vk
-            .iter()
-            .cloned()
-            .zip(&AHPForR1CS::<F>::INDEXER_POLYNOMIALS)
-            .map(|(c, l)| LabeledCommitment::new(l.to_string(), c, None))
-            .chain(first_comms.into_iter())
-            .chain(second_comms.into_iter())
-            .chain(third_comms.into_iter())
-            .collect();
-
-        // Gather commitment randomness together.
-        let comm_rands: Vec<PC::Randomness> = index_pk
-            .index_comm_rands
-            .clone()
-            .into_iter()
-            .chain(first_comm_rands)
-            .chain(second_comm_rands)
-            .chain(third_comm_rands)
-            .collect();
-
-        // Compute the AHP verifier's query set.
-        let (query_set, verifier_state) = AHPForR1CS::verifier_query_set(verifier_state, &mut fs_rng);
-        let lc_s = AHPForR1CS::construct_linear_combinations(&public_input, &polynomials, &verifier_state)?;
-
-        let eval_time = start_timer!(|| "Evaluating linear combinations over query set");
-        let mut evaluations = Vec::new();
-        for (label, point) in &query_set {
-            let lc = lc_s
-                .iter()
-                .find(|lc| &lc.label == label)
-                .ok_or_else(|| ahp::Error::MissingEval(label.to_string()))?;
-            let eval = polynomials.get_lc_eval(&lc, *point)?;
-            if !AHPForR1CS::<F>::LC_WITH_ZERO_EVAL.contains(&lc.label.as_ref()) {
-                evaluations.push(eval);
-            }
-        }
-        end_timer!(eval_time);
-
-        fs_rng.absorb(&evaluations);
-        let opening_challenge: F = u128::rand(&mut fs_rng).into();
-
-        let pc_proof = PC::open_combinations(
-            &index_pk.committer_key,
-            &lc_s,
-            polynomials,
-            &labeled_comms,
-            &query_set,
-            opening_challenge,
-            &comm_rands,
-            Some(zk_rng),
-        )
-        .map_err(Error::from_pc_err)?;
-
-        // Gather prover messages together.
-        let prover_messages = vec![prover_first_msg, prover_second_msg, prover_third_msg];
-
-        let proof = Proof::new(commitments, evaluations, prover_messages, pc_proof);
-        proof.print_size_info();
-        end_timer!(prover_time);
-        Ok(proof)
-    }
-
-    /// Verify that a proof for the constrain system defined by `C` asserts that
-    /// all constraints are satisfied.
-    pub fn verify<R: RngCore>(
-        index_vk: &IndexVerifierKey<F, PC>,
-        public_input: &[F],
-        proof: &Proof<F, PC>,
-        rng: &mut R,
-    ) -> Result<bool, Error<PC::Error>> {
-        let verifier_time = start_timer!(|| "Marlin::Verify");
-
-        let mut fs_rng =
-            FiatShamirRng::<D>::from_seed(&to_bytes![&Self::PROTOCOL_NAME, &index_vk, &public_input].unwrap());
-
-        // --------------------------------------------------------------------
-        // First round
-
-        let first_comms = &proof.commitments[0];
-        fs_rng.absorb(&to_bytes![first_comms, proof.prover_messages[0]].unwrap());
-
-        let (_, verifier_state) = AHPForR1CS::verifier_first_round(index_vk.index_info, &mut fs_rng)?;
-        // --------------------------------------------------------------------
-
-        // --------------------------------------------------------------------
-        // Second round
-        let second_comms = &proof.commitments[1];
-        fs_rng.absorb(&to_bytes![second_comms, proof.prover_messages[1]].unwrap());
-
-        let (_, verifier_state) = AHPForR1CS::verifier_second_round(verifier_state, &mut fs_rng);
-        // --------------------------------------------------------------------
-
-        // --------------------------------------------------------------------
-        // Third round
-        let third_comms = &proof.commitments[2];
-        fs_rng.absorb(&to_bytes![third_comms, proof.prover_messages[2]].unwrap());
-
-        let verifier_state = AHPForR1CS::verifier_third_round(verifier_state, &mut fs_rng);
-        // --------------------------------------------------------------------
-
-        // Collect degree bounds for commitments. Indexed polynomials have *no*
-        // degree bounds because we know the committed index polynomial has the
-        // correct degree.
-        let index_info = index_vk.index_info;
-        let degree_bounds = vec![None; index_vk.index_comms.len()]
-            .into_iter()
-            .chain(AHPForR1CS::prover_first_round_degree_bounds(&index_info))
-            .chain(AHPForR1CS::prover_second_round_degree_bounds(&index_info))
-            .chain(AHPForR1CS::prover_third_round_degree_bounds(&index_info));
-
-        // Gather commitments in one vector.
-        let commitments = index_vk
-            .iter()
-            .chain(first_comms)
-            .chain(second_comms)
-            .chain(third_comms)
-            .cloned()
-            .zip(AHPForR1CS::<F>::polynomial_labels())
-            .zip(degree_bounds)
-            .map(|((c, l), d)| LabeledCommitment::new(l, c, d));
-
-        let (query_set, verifier_state) = AHPForR1CS::verifier_query_set(verifier_state, &mut fs_rng);
-
-        fs_rng.absorb(&proof.evaluations);
-        let opening_challenge: F = u128::rand(&mut fs_rng).into();
-
-        let mut evaluations = Evaluations::new();
-        let mut proof_evals = proof.evaluations.iter();
-        for q in query_set.iter().cloned() {
-            if AHPForR1CS::<F>::LC_WITH_ZERO_EVAL.contains(&q.0.as_ref()) {
-                evaluations.insert(q, F::zero());
-            } else {
-                evaluations.insert(q, *proof_evals.next().unwrap());
-            }
-        }
-
-        let lc_s = AHPForR1CS::construct_linear_combinations(&public_input, &evaluations, &verifier_state)?;
-
-        let evaluations_are_correct = PC::check_combinations(
-            &index_vk.verifier_key,
-            &lc_s,
-            commitments,
-            &query_set,
-            &evaluations,
-            &proof.pc_proof,
-            opening_challenge,
-            rng,
-        )
-        .map_err(Error::from_pc_err)?;
-
-        if !evaluations_are_correct {
-            eprintln!("PC::Check failed");
-        }
-        end_timer!(verifier_time, || format!(
-            " PC::Check for AHP Verifier linear equations: {}",
-            evaluations_are_correct
-        ));
-        Ok(evaluations_are_correct)
-    }
-}
+// /// A structured reference string which will be used to derive a circuit-specific
+// /// common reference string
+// pub type SRS<E> = crate::UniversalSRS<<E as PairingEngine>::Fr, MultiPC<E>>;
+//
+// /// A circuit-specific proving key.
+// pub type ProverKey<E> = crate::CircuitProvingKey<<E as PairingEngine>::Fr, MultiPC<E>>;
+//
+// /// A circuit-specific verification key.
+// pub type VerifierKey<E> = crate::CircuitVerifyingKey<<E as PairingEngine>::Fr, MultiPC<E>>;
+//
+// impl<E: PairingEngine> From<Parameters<E>> for VerifierKey<E> {
+//     fn from(params: Parameters<E>) -> Self {
+//         params.verifier_key
+//     }
+// }
